requirements.txt
json
ast
requests
os
Pillow
keras
scikit-learn
numpy
glob
h5py
pickle